Removing comms
