--- comms/.gitignore
@@ -1,109 +1,0 @@
-# Prerequisites
-*.d
-
-# Compiled Object files
-*.slo
-*.lo
-*.o
-*.obj
-
-# Precompiled Headers
-*.gch
-*.pch
-
-# Compiled Dynamic libraries
-*.so
-*.dylib
-*.dll
-
-# Fortran module files
-*.mod
-*.smod
-
-# Compiled Static libraries
-*.lai
-*.la
-*.a
-*.lib
-
-# Executables
-*.exe
-*.out
-*.app
-
-# Prerequisites
-*.d
-
-# Compiled Object files
-*.slo
-*.lo
-*.o
-*.obj
-
-# Precompiled Headers
-*.gch
-*.pch
-
-# Compiled Dynamic libraries
-*.so
-*.dylib
-*.dll
-
-# Fortran module files
-*.mod
-*.smod
-
-# Compiled Static libraries
-*.lai
-*.la
-*.a
-*.lib
-
-# Executables
-*.exe
-*.out
-*.app
-
-# Folders
-build/
-/testing-libgdf
-gtests/
-
-# Eclipse generated files by cmake
-.cproject
-.project
-
-# CMake generated files
-LastTest.log
-cmake_install.cmake
-CMakeCache.txt
-CTestTestfile.cmake
-Makefile
-CTestCostData.txt
-DartConfiguration.tcl
-Testing/
-*DartConfiguration.tcl
-CMakeCache.txt
-CMakeFiles/
-Makefile
-cmake_install.cmake
-
-\.vscode/
-
-#generated folders (in source)
-blazingdb-protocol_ep-prefix
-google-flatbuffers_ep-prefix
-
-\.idea/
-
-resources/tpch-generator/__pycache__/
-
-#Qt Creator files
-CMakeLists.txt.user
-
-docs/html/
-docs.tar.gz
-docs.zip
-docs.bz2
-
-cmake-build-debug/
