--- comms/README.md
@@ -1,12 +1,0 @@
-# <div align="left"><img src="img/blazingSQL.png" width="200px"/>&nbsp;BlazingSQL</div>
-
-# BlazingSQL Communication Library
-
-The BlazingSQL Communication Library is a C++ library that contains all the logic for worker to worker communication
-
-For more information about the BlazingSQL engine please see our [main repo](https://github.com/BlazingDB/pyBlazing). There you will find the following information:
-- Getting started info
-- How to use BlazingSQL
-- Build instructions
-- Contribution Guide
-- How to file an issue
