+++ engine/src/CalciteInterpreter.h
@@ -4,7 +4,7 @@
 #include <iostream>
 #include <string>
 #include <vector>
-#include <blazingdb/manager/Context.h>
+#include <execution_graph//Context.h>
 #include "execution_graph/logic_controllers/taskflow/graph.h"
 
 using blazingdb::manager::Context;
