+++ engine/src/communication/CommunicationInterface/messageListener.hpp
@@ -4,7 +4,7 @@
 #include <map>
 
 #include "utilities/ctpl_stl.h"
-#include "blazingdb/concurrency/BlazingThread.h"
+#include "ExceptionHandling/BlazingThread.h"
 #include <ucp/api/ucp.h>
 #include <ucp/api/ucp_def.h>
 #include "messageReceiver.hpp"
