+++ engine/src/communication/CommunicationInterface/messageReceiver.hpp
@@ -6,7 +6,7 @@
 #include <tuple>
 #include <memory>
 #include <rmm/device_buffer.hpp>
-#include <blazingdb/transport/ColumnTransport.h>
+#include <transport/ColumnTransport.h>
 
 #include "serializer.hpp"
 #include "execution_graph/logic_controllers/CacheMachine.h"
