+++ engine/src/communication/CommunicationInterface/protocols.cpp
@@ -11,7 +11,7 @@
 
 
 #include "execution_graph/logic_controllers/CacheMachine.h"
-#include "blazingdb/transport/io/reader_writer.h"
+#include "transport/io/reader_writer.h"
 
 constexpr size_t NUMBER_RETRIES = 20;
 constexpr size_t FILE_RETRY_DELAY = 20;
