+++ engine/src/communication/messages/GPUComponentMessage.h
@@ -5,8 +5,8 @@
 #include <spdlog/sinks/basic_file_sink.h>
 #include <spdlog/sinks/stdout_color_sinks.h>
 
-#include <blazingdb/transport/ColumnTransport.h>
-#include <blazingdb/transport/Node.h>
+#include <transport/ColumnTransport.h>
+#include <transport/Node.h>
 #include <communication/messages/MessageUtil.cuh>
 
 #include <map>
