+++ engine/src/execution_graph/logic_controllers/BlazingHostTable.cpp
@@ -1,5 +1,5 @@
 #include "BlazingHostTable.h"
-#include <blazingdb/transport/ColumnTransport.h>
+#include <transport/ColumnTransport.h>
 
 #include <vector>
 #include <string>
