+++ engine/src/execution_graph/logic_controllers/LogicalProject.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <blazingdb/manager/Context.h>
+#include <execution_graph/Context.h>
 
 #include "LogicPrimitives.h"
 #include "execution_graph/logic_controllers/BlazingColumn.h"
