+++ engine/src/operators/GroupBy.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <blazingdb/manager/Context.h>
+#include <execution_graph//Context.h>
 #include <string>
 #include <vector>
 #include <tuple>
